refactor(editor): complete state and output extractions
